chore : 2차 피드백 수용

+++ winwin-be-api/src/main/java/com/dpm/winwin/api/member/service/MemberQueryService.java
@@ -36,7 +36,7 @@ public class MemberQueryService {
         MemberReadResponse memberReadResponse =  memberRepository.readMemberInfo(memberId)
                 .orElseThrow(() -> new BusinessException(MEMBER_NOT_FOUND));
 
-        BigDecimal likeCount = member.getLikeCount();
+        BigDecimal likeCount = BigDecimal.valueOf(member.getLikeCount());
         String likeCounts = likeCount.toString();
